ml_pipeline/requirements.txt
# Core ML Libraries (optimized for Apple Silicon)
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Gradient Boosting Models
xgboost>=2.0.0  # Apple Silicon optimized
lightgbm>=4.0.0  # Apple Silicon optimized
catboost>=1.2.0

# Model Export
onnx>=1.14.0
skl2onnx>=1.16.0
onnxruntime>=1.16.0

# Hyperparameter Tuning
optuna>=3.3.0

# Model Explainability
shap>=0.43.0

# Data Processing
pyarrow>=13.0.0  # For Parquet files

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Progress Bars
tqdm>=4.66.0

# Configuration
pyyaml>=6.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
joblib>=1.3.0
